Fix sign-compare build error
A recent change introduced a compare between 'int ret' and
'array.size()'. array.size() is evaluating to an unsigned int which is
causing a sign-compare build error.

This adds a static_cast to remove the sign-compare difference.

Tested:
Confirmed that it builds without a sign-compare error.

+++ src/fru_device.cpp
@@ -254,7 +254,7 @@ static int i2cSmbusWriteThenRead(int file, uint16_t address,
 
     int ret = ioctl(file, I2C_RDWR, &rdwr);
 
-    return (ret == msgs.size()) ? msgs[1].len : -1;
+    return (ret == static_cast<int>(msgs.size())) ? msgs[1].len : -1;
 }
 
 static int64_t readBlockData(bool is16bit, int file, uint16_t address,
